requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.3.0
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==19.7.1
Automat==20.2.0
autopep8==1.5.4
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography
daphne==2.5.0
Django==2.2.16
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-rest-swagger==2.2.0
django-solo==1.1.3
django-upload-validator==1.0.2
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
drf-yasg==1.17.0
et-xmlfile==1.0.1
Fabric3==1.14.post1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.3.1
isort==5.6.4
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
openapi-codec==1.3.2
openpyxl==2.6.2
packaging==19.2
paramiko==2.6.0
Pillow==7.0.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.3.0
pyOpenSSL==23.0.0
pyparsing==2.4.2
python-magic==0.4.15
pytz==2019.1
redis==3.5.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.2
tqdm==4.50.2
Twisted==19.2.1
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.2
pydantic
